% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_valid_split}
\alias{check_valid_split}
\title{Check if a node can be split while respecting minimum size}
\usage{
check_valid_split(hc, node_id, min_size)
}
\arguments{
\item{hc}{hclust object}

\item{node_id}{Node to potentially split}

\item{min_size}{Minimum cells per resulting subcluster}
}
\value{
List with can_split, height, left_id, right_id, left_size, right_size
}
\description{
Check if a node can be split while respecting minimum size
}
\keyword{internal}
