% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_node_members}
\alias{get_node_members}
\title{Get all cell indices under a dendrogram node}
\usage{
get_node_members(hc, node_id)
}
\arguments{
\item{hc}{hclust object}

\item{node_id}{Node identifier}
}
\value{
Vector of cell indices (positions in hc$labels)
}
\description{
Get all cell indices under a dendrogram node
}
\keyword{internal}
